Try to use a task dialog instead of a message box
This is with the hope that a task dialog will:

- Allow the text to be selected. Pressing Ctrl+C with a message box
  up actually does copy its contents, with the parts separated by
  horizontal dashed lines, to the clipboard. But this is not well
  known, and also not really what one ever wants: one wants to be
  able to copy the specific contents.

- Play better with display scaling. I was able to make the message
  box DPI aware enough to not be blurry, but it was not actually
  scaled up, so it still appeared too small.

I am unsure if this will fix both problems, but I hope so.

This is not done yet, and currently always fails, because it does
not take the necessary extra steps to use COM controls in a program
that targets the console subsystem. Currently it fails with:

error: process didn't exit successfully: `target\debug\showargsw.exe "a b" c` (exit code: 0xc0000139, STATUS_ENTRYPOINT_NOT_FOUND)